move stories to docs/components
